fix(nvim): fix border size for LSP signature help popup
When you press 'K' to get LSP signature help on a function object, it
will create a very big popup window if there are many parameters in the
function object and/or the documentation is long. The problem is caused
by the noice plugin. Fixed this by reducing the default maxheight and
maxwidth of the hover window.